+++ doc_test.go
@@ -6,4 +6,10 @@ func TestVersionNonEmpty(t *testing.T) {
 	if Version == "" {
 		t.Fatal("Version must not be empty")
 	}
+	if Commit == "" {
+		t.Fatal("Commit must not be empty")
+	}
+	if BuildDate == "" {
+		t.Fatal("BuildDate must not be empty")
+	}
 }